Session 30- Assignment30_1
4. Associated Data Files
 Use “Data.xlsx”
5. Problem Statement
 • Using the excel spreadsheet “Data.xlsx”, create a series of Pivot Tables (or Pivot Charts) to answer the following questions.
• What are the averages for purchases in each region?  
• What form of payment is most common? 
 • Do our customers shop at work (most likely between 8:00 and 17:00) or at home? 
 • Turn in a printout of the pivot tables you used to answer each question.
Note: To answer number 3, a new column is created in the spreadsheet labeled "Hour of Day" and set it to just the hour instead of hour and minutes. 
In other words =hour(G8). The students who are excel power users should use an "if" formula to create a value of either day (time between 8:00 and 17:00) or night (not between 8:00 and 17:00).